move CommandBroker, CommandHistory to commands.py
