+++ sync_threads/commands.py
@@ -1,4 +1,5 @@
 import os
+from queue import Queue
 
 
 CommandNotFoundError = KeyError  # TODO: add implementation
@@ -19,6 +20,33 @@ class Commands:
             return command_func(*args)
         except KeyError:
             raise CommandNotFoundError
+
+
+class CommandBroker:
+    def __init__(self):
+        self.__listeners = set()
+
+    def subscribe(self, listener):
+        self.__listeners.add(listener)
+
+    def unsubscribe(self, listener):
+        self.__listeners.remove(listener)
+
+    def notify(self, command):
+        for listener in self.__listeners:
+            listener.update(command)
+
+
+class CommandHistory:
+    def __init__(self, history_size: int = 100):
+        self.history = Queue(maxsize=history_size)
+
+    def update(self, command):
+        self.history.put(command)
+
+    def history_items(self, last_items=0):
+        history_items = list(self.history.queue)
+        return history_items[-last_items:]
 
 
 @Commands.register
